+++ src/main/java/com/mechanical_workshop_usm/mechanic_info_module/dto/CreateMechanicResponse.java
@@ -3,6 +3,6 @@ package com.mechanical_workshop_usm.mechanic_info_module.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record CreateMechanicResponse(
-        long id,
+        int id,
         @JsonProperty("registration_number") String registrationNumber
 ) {}
