+++ src/main/java/com/mechanical_workshop_usm/record_module/record/dto/CreateRecordResponse.java
@@ -5,6 +5,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public record CreateRecordResponse(
         @JsonProperty("id") int id,
         @JsonProperty("car_id") int carId,
-        @JsonProperty("client_info_id") long clientInfoId,
+        @JsonProperty("client_info_id") int clientInfoId,
         @JsonProperty("mechanic_info_id") int mechanicInfoId
 ) {}
